+++ lib/presentation/screens/home/tabs/statistics_tab/widgets/legend_widget.dart
